+++ src/Mirakl/MMP/Common/Domain/Message/Thread/ThreadMetadata.php
@@ -4,10 +4,19 @@ namespace Mirakl\MMP\Common\Domain\Message\Thread;
 use Mirakl\Core\Domain\MiraklObject;
 
 /**
- * @method  int         getTotalCount()
- * @method  $this       setTotalCount(string $totalCount)
- * @method  \DateTime   getLastMessageDate()
- * @method  $this       setLastMessageDate(\DateTime $lastMessageDate)
+ * @method int          getTotalCount()
+ * @method $this        setTotalCount(int $totalCount)
+ * @method \DateTime    getLastMessageDate()
+ * @method $this        setLastMessageDate(\DateTime $lastMessageDate)
+ * @method ThreadSender getLastSender()
+ * @method $this        setLastSender(ThreadSender $threadSender)
  */
 class ThreadMetadata extends MiraklObject
-{}
+{
+    /**
+     * @var array
+     */
+    protected static $dataTypes = [
+        'last_sender' => [ThreadSender::class, 'create'],
+    ];
+}