+++ src/Mirakl/MMP/Common/Domain/Offer/AbstractOffer.php
@@ -74,6 +74,8 @@ use Mirakl\MMP\Common\Domain\Offer\Price\OfferPricing;
  * @method  $this                                 setTotalPrice(float $totalPrice)
  * @method  ShippingPriceByZoneAndTypeCollection  getShippingTypes()
  * @method  $this                                 setShippingTypes(ShippingPriceByZoneAndTypeCollection $shippingPricesByZoneAndType)
+ * @method  \DateTime                             getShippingDeadline()
+ * @method  $this                                 setShippingDeadline(\DateTime $shippingDeadline)
  */
 abstract class AbstractOffer extends MiraklObject
 {