+++ src/Mirakl/MMP/Shop/Domain/Payment/Transaction/TransactionOrderTax.php
@@ -6,6 +6,8 @@ use Mirakl\Core\Domain\MiraklObject;
 /**
  * @method  string  getCode()
  * @method  $this   setCode(string $code)
+ * @method  float   getRate()
+ * @method  $this   setRate(float $rate)
  */
 class TransactionOrderTax extends MiraklObject
 {}